Updates instructions to reflect gif example
Closes #3  
Hey @bbaneydev! Thanks for flagging this! I went ahead and added some instructions in the readme to reflect the `<button>` element's text changing based on state, like how it is in the gif. Please note these changes might not be reflected in your cohort's Canvas, but are reflected in the github repository, as well as in future cohorts Canvas pages.
We appreciate you raising this issue! -Gracie